Commit for message total gathering and storage by day for channels-teams

+++ src/main/java/edu/asu/cassess/persist/repo/slack/SlackMessageRepo.java
@@ -3,5 +3,5 @@ package edu.asu.cassess.persist.repo.slack;
 import edu.asu.cassess.persist.entity.slack.SlackMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SlackMessageRepo extends JpaRepository<SlackMessage, Long> {
+public interface SlackMessageRepo extends JpaRepository<SlackMessage, Integer> {
 }
